more deprecations
